Filter out loopback ip from /etc/hosts
I added a condition to filter out loopback address from /etc/hosts, because I have a lot of websites that I want to block and I map it to `127.0.0.1` for all of them in my host file. It might be useful for other people who do the same thing as me. So feel free to accept or reject. Overall your code is very useful. Thanks for developing it. 👍